# artifacts/api-server/src/routes/auth.ts
import { Router, type IRouter } from "express";
import bcrypt from "bcryptjs";
import { db, usersTable } from "@workspace/db";
import { eq } from "drizzle-orm";
import { z } from "zod";
import {
  clearSessionCookie,
  loadUserFromCookie,
  setSessionCookie,
  toAuthUser,
} from "../lib/auth";
import { isAllowedOrigin } from "../lib/cors";

const router: IRouter = Router();

const LoginBody = z.object({
  email: z.string().email(),
  password: z.string().min(1),
});

router.post("/auth/login", async (req, res): Promise<void> => {
  const origin = req.headers["origin"] as string | undefined;
  if (origin !== undefined && !isAllowedOrigin(origin)) {
    res.status(403).json({ error: "Forbidden" });
    return;
  }
  const parsed = LoginBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: "Invalid email or password" });
    return;
  }
  const email = parsed.data.email.toLowerCase().trim();
  const [user] = await db
    .select()
    .from(usersTable)
    .where(eq(usersTable.email, email));
  if (!user || !user.isActive) {
    res.status(401).json({ error: "Invalid email or password" });
    return;
  }
  // The automation service account authenticates only via the
  // INTEGRATION_API_KEY bearer token. Reject password login attempts
  // even if a hash somehow matched, so a leaked password cannot
  // bypass the bearer-only contract.
  if (user.role === "service") {
    res.status(401).json({ error: "Invalid email or password" });
    return;
  }
  const ok = await bcrypt.compare(parsed.data.password, user.passwordHash);
  if (!ok) {
    res.status(401).json({ error: "Invalid email or password" });
    return;
  }
  setSessionCookie(res, user.id);
  res.json({ user: toAuthUser(user) });
});

router.post("/auth/logout", (_req, res): void => {
  clearSessionCookie(res);
  res.json({ ok: true });
});

router.get("/auth/me", async (req, res): Promise<void> => {
  const user = await loadUserFromCookie(req);
  if (!user) {
    res.status(401).json({ error: "Not authenticated" });
    return;
  }
  res.json({ user });
});

export default router;