# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Life House Reentry Finance Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserResponse"
...
  /credits:
    get:
      operationId: listCredits
      tags: [credits]
      summary: List credits, optionally filtered by status
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/CreditStatus"
      responses:
        "200":
          description: Credits list
          content:
            application/json:
              schema:
                type: object
                properties:
                  credits:
                    type: array
                    items:
                      $ref: "#/components/schemas/Credit"
                required: [credits]
    post:
      operationId: createCredit
      tags: [credits]
      summary: Create a new credit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCreditBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  credit:
                    $ref: "#/components/schemas/Credit"
                required: [credit]
  /credits/{id}:
    put:
      operationId: updateCredit
      tags: [credits]
      summary: Update an existing credit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCreditBody"
      responses:
        "200":
          description: Updated credit
          content:
            application/json:
              schema:
                type: object
                properties:
                  credit:
                    $ref: "#/components/schemas/Credit"
                required: [credit]
    delete:
      operationId: deleteCredit
      tags: [credits]
      summary: Delete a credit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required: [ok]
...
components:
  schemas:
...
    CreditStatus:
      type: string
      enum: [pipeline, received, delayed, write-off, opportunity]
    Credit:
      type: object
      properties:
        id:
          type: integer
        source:
          type: string
        programId:
          type: integer
          nullable: true
        programName:
          type: string
          nullable: true
        amount:
          type: number
        expectedDate:
          type: string
          nullable: true
        receivedDate:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/CreditStatus"
        notes:
          type: string
          nullable: true
        submittedBy:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - source
        - programId
        - programName
        - amount
        - expectedDate
        - receivedDate
        - status
        - notes
        - submittedBy
        - createdAt
...
    CreateCreditBody:
      type: object
      properties:
        source:
          type: string
        programId:
          type: integer
          nullable: true
        amount:
          type: number
        expectedDate:
          type: string
          nullable: true
        receivedDate:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/CreditStatus"
        notes:
          type: string
          nullable: true
        submittedBy:
          type: string
          nullable: true
      required:
        - source
        - amount
...
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          # `service` is the automation account authenticated via
          # INTEGRATION_API_KEY bearer token. It is hidden from the
          # admin user list and never returned by /auth/me or /auth/login,
          # but is included here so the type union stays consistent
          # with the server.
          enum: [admin, approver, submitter, service]
      required:
        - id
        - email
        - firstName
        - lastName
        - role
    AuthUserResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/AuthUser"
      required: [user]
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
      required: [email, password]
...